Large structural changes to the WikiData classes to reduce redundant code.
We now define a BaseWikiData class from which all backends inherit.

The compact_sqlite database format was updated to bring it in line with the
original_sqlite format.

This commit also adds a new wiki specific option

"caseinsensitivewikiwords" = [True/False]

Which allows wikiwords to be case independent (will not work with gadfly
backend)